backend/requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
motor==3.7.1
python-dotenv==1.0.1
pymongo==4.16.0
firebase-admin==6.8.0
twilio==9.10.4
groq==1.1.2
httpx==0.27.0
apscheduler==3.11.2
pydantic==2.12.5